Adjust tests to new module structure
