+++ tests/apis/binance/general.rs
