+++ tests/apis/coinbase/websocket.rs
