+++ tests/apis/mod.rs
@@ -1,0 +1,2 @@
+mod binance;
+mod coinbase;
