--- tests/openlimits/mod.rs
@@ -1,9 +1,0 @@
-mod binance {
-    mod account;
-    mod market;
-}
-
-mod coinbase {
-    mod account;
-    mod market;
-}
